api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-bigquery==3.15.0
google-cloud-storage==2.14.0
pydantic==2.5.0
python-multipart==0.0.6

